requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.6.0
billiard==3.6.3.0
celery==4.4.2
certifi==2018.11.29
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.3
django-elasticsearch-dsl==0.5.1
django-phonenumber-field==2.3.1
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
environs==7.4.0
gunicorn==20.0.4
humanize==1.0.0
importlib-metadata==1.6.0
ipaddress==1.0.22
kombu==4.6.8
marshmallow==3.6.0
mock==2.0.0
munch==2.5.0
pbr==5.1.2
phonenumbers==8.12.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.13.0
python-http-client==3.2.6
pytz==2017.3
redis==3.5.0
sendgrid==6.2.0
six==1.11.0
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0

PyJWT~=1.7.1
djangorestframework~=3.11.1